+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 7.1.0 (2021-09-20)
+
+* Custom Dexguard paths are resolved correctly (relative to the project)
+  [#421](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/421)
+
+* Support extracting SO files from bugsnag AARs for library modules
+  [#423](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/423)
+
+* Support DexGuard with AGP 7
+  [#426](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/426)
+
 ## 7.0.0 (2021-08-12)
 
 * Address task dependency warning when using APK splits